Moved all content that was previously in the center section of the MainWindowView's borderpane to its own class
