+++ src/main/resources/view/main-window.fxml
@@ -1,103 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ButtonBar?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.SplitPane?>
-<?import javafx.scene.control.Tab?>
-<?import javafx.scene.control.TabPane?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextArea?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
 
-<AnchorPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="1050.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
+<AnchorPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="1050.0"
+            xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <BorderPane fx:id="mainWindowLayout" layoutY="27.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-            <center>
-                <SplitPane fx:id="mainViewSplit" dividerPositions="0.7" orientation="VERTICAL" BorderPane.alignment="CENTER">
-                    <items>
-                        <VBox>
-                            <children>
-                                <ScrollPane fitToHeight="true" fitToWidth="true" VBox.vgrow="ALWAYS">
-                                    <content>
-                                        <TableView fx:id="modTable" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308">
-                                            <placeholder>
-                                                <Label text="No Mods Installed" />
-                                            </placeholder>
-                                            <columns>
-                                                <TableColumn fx:id="modName" prefWidth="200.0" text="Mod Name" />
-                                                <TableColumn fx:id="loadPriority" prefWidth="85.0" text="Priority" />
-                                                <TableColumn fx:id="modLastUpdated" prefWidth="115.0" text="Last Updated" />
-                                                <TableColumn fx:id="modVersion" prefWidth="85.0" text="Version" />
-                                                <TableColumn fx:id="modSource" prefWidth="85.0" text="Import Source" visible="false" />
-                                                <TableColumn fx:id="modType" prefWidth="100.0" text="Source" />
-                                                <TableColumn fx:id="modCategory" prefWidth="125.0" text="Category/Tags" />
-                                            </columns>
-                                            <columnResizePolicy>
-                                                <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-                                            </columnResizePolicy>
-                                        </TableView>
-                                    </content>
-                                </ScrollPane>
-                                <HBox fx:id="tableActions" alignment="CENTER_LEFT" spacing="10.0" VBox.vgrow="NEVER">
-                                    <children>
-                                        <HBox alignment="CENTER_LEFT" spacing="10.0">
-                                            <children>
-                                                <ComboBox fx:id="modImportDropdown" prefWidth="250.0" promptText="Add mods from..." />
-                                            </children>
-                                        </HBox>
-                                        <ButtonBar buttonOrder="U+FBIX_NCYOA_R">
-                                            <buttons>
-                                                <Button fx:id="importModlist" mnemonicParsing="false" text="Import Modlist" />
-                                                <Button fx:id="exportModlist" mnemonicParsing="false" text="Export Modlist" />
-                                                <Button fx:id="resetModlist" mnemonicParsing="false" text="Reset Modlist" />
-                                                <Button fx:id="injectModlist" mnemonicParsing="false" onAction="#applyModlist" text="Apply Modlist" />
-                                                <Button fx:id="launchSpaceEngineers" mnemonicParsing="false" onAction="#launchSpaceEngineers" text="Launch SE" />
-                                            </buttons>
-                                        </ButtonBar>
-                                        <HBox alignment="CENTER_LEFT" spacing="10.0" />
-                                    </children>
-                                    <VBox.margin>
-                                        <Insets />
-                                    </VBox.margin>
-                                    <padding>
-                                        <Insets bottom="5.0" left="10.0" right="10.0" top="5.0" />
-                                    </padding>
-                                </HBox>
-                            </children>
-                        </VBox>
-                        <TabPane fx:id="informationPane" tabDragPolicy="REORDER">
-                            <tabs>
-                                <Tab fx:id="modDescriptionTab" onClosed="#closeModDescriptionTab" text="Mod Description">
-                                    <content>
-                                        <ScrollPane fitToHeight="true" fitToWidth="true" hbarPolicy="ALWAYS" maxWidth="1.7976931348623157E308">
-                                            <content>
-                                                <TextArea editable="false" />
-                                            </content>
-                                        </ScrollPane>
-                                    </content>
-                                </Tab>
-                                <Tab fx:id="logTab" onClosed="#closeLogTab" text="Log">
-                                    <content>
-                                        <ListView fx:id="viewableLog" focusTraversable="false" />
-                                    </content>
-                                </Tab>
-                            </tabs>
-                            <padding>
-                                <Insets bottom="5.0" />
-                            </padding>
-                        </TabPane>
-                    </items>
-                </SplitPane>
-            </center>
+        <BorderPane fx:id="mainWindowLayout" layoutY="27.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
+                    AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
         </BorderPane>
     </children>
 </AnchorPane>
